refactor: 체크아웃 페이지 정리
1차 배포용
